+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## [*Unreleased*] - <*2*>
+
+### Changes
+
+* Upgraded `@microsoft/teams-js` SDK to version 2
 ## [*3.1.1*] - <*2021-09-30*>
 
 ### Changes